requirements.txt
# X (Twitter) API
tweepy==4.14.0

# Web framework
flask==3.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Time parsing
parsedatetime==2.6
python-dateutil==2.8.2

# Scheduling
APScheduler==3.10.4

# Environment variables
python-dotenv==1.0.0

# Utilities
pytz==2023.3

